+++ netCoreMsal/Program.cs
@@ -10,22 +10,36 @@ using System.IO;
 using System.Text;
 using System.Text.Json;
 using System.Linq;
+using System.Security;
 using Microsoft.Identity.Client;
 
 namespace netCoreMsal
 {
     internal class Program
     {
+        private AuthenticationResult authenticationResult = default;
         private string clientId = "1950a258-227b-4e31-a9cf-717495945fc2";
+        private string clientName = null;
+        private string clientSecret = null;
+        private IConfidentialClientApplication confidentialClientApp;
+        private List<string> defaultScope = new List<string>() { ".default" };
         private bool detail = false;
         private bool help = false;
         private IPublicClientApplication publicClientApp;
         private string redirectUri = null; //"http://localhost";
         private string resource = null;
-        private List<string> scopes = new List<string>() { ".default" };
+        private List<string> scopes = new List<string>();
         private string tenantId = "common";
 
-        static void Main(string[] args)
+        public void MsalLoggerCallback(LogLevel level, string message, bool containsPII)
+        {
+            if (detail)
+            {
+                Console.WriteLine($"// {level} {message}");
+            }
+        }
+
+        private static void Main(string[] args)
         {
             try
             {
@@ -37,9 +51,81 @@ namespace netCoreMsal
             }
         }
 
-        public void MsalLoggerCallback(LogLevel level, string message, bool containsPII)
+        private void Authorize()
         {
-            Console.WriteLine($"//{DateTime.Now} {level} {containsPII} {message}");
+            try
+            {
+                List<string> newScopes = new List<string>();
+
+                foreach (string scope in scopes)
+                {
+                    newScopes.Add($"{resource}/{scope}");
+                }
+
+                scopes = newScopes;
+
+                if (string.IsNullOrEmpty(clientSecret))
+                {
+                    // user creds 
+                    publicClientApp = PublicClientApplicationBuilder
+                        .Create(clientId)
+                        .WithAuthority(AzureCloudInstance.AzurePublic, tenantId)
+                        .WithLogging(MsalLoggerCallback, LogLevel.Verbose, true, true)
+                        .WithDefaultRedirectUri()
+                        .Build();
+
+                    TokenCacheHelper.EnableSerialization(publicClientApp.UserTokenCache);
+                    authenticationResult = publicClientApp
+                        .AcquireTokenSilent(defaultScope, publicClientApp.GetAccountsAsync().Result.FirstOrDefault())
+                        .ExecuteAsync().Result;
+                }
+                else
+                {
+                    // client creds
+                    if (string.IsNullOrEmpty(clientName))
+                    {
+                        clientName = clientId;
+                    }
+
+                    confidentialClientApp = ConfidentialClientApplicationBuilder
+                        .CreateWithApplicationOptions(new ConfidentialClientApplicationOptions
+                        {
+                            ClientId = clientId,
+                            RedirectUri = redirectUri,
+                            ClientSecret = clientSecret,
+                            TenantId = tenantId,
+                            ClientName = clientName
+                        })
+                        .WithAuthority(AzureCloudInstance.AzurePublic, tenantId)
+                        .WithLogging(MsalLoggerCallback, LogLevel.Verbose, true, true)
+                        .Build();
+
+                    TokenCacheHelper.EnableSerialization(confidentialClientApp.UserTokenCache);
+                    authenticationResult = confidentialClientApp
+                        .AcquireTokenForClient(scopes.Count > 0 ? scopes : defaultScope)
+                        .ExecuteAsync().Result;
+                }
+            }
+            catch (MsalUiRequiredException)
+            {
+                authenticationResult = publicClientApp
+                    .AcquireTokenInteractive(defaultScope)
+                    .ExecuteAsync().Result;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine($"{e}");
+            }
+
+            if (scopes.Count > 0)
+            {
+                authenticationResult = publicClientApp
+                    .AcquireTokenSilent(scopes, publicClientApp.GetAccountsAsync().Result.FirstOrDefault())
+                    .ExecuteAsync().Result;
+            }
+
+            //Console.WriteLine($"{JsonSerializer.Serialize(authenticationResult)}");
+            FormatJsonOutput(authenticationResult);
         }
 
         private void Execute(string[] args)
@@ -52,10 +138,12 @@ namespace netCoreMsal
                 if (arg == "resource") { resource = args[++i]; }
                 if (arg == "redirecturi") { redirectUri = args[++i]; }
                 if (arg == "clientid") { clientId = args[++i]; }
+                if (arg == "clientname") { clientName = args[++i]; }
+                if (arg == "clientsecret") { clientSecret = args[++i]; }
                 if (arg == "tenantid") { tenantId = args[++i]; }
                 if (arg == "detail") { detail = true; }
                 if (arg == "?") { help = true; }
-                if (arg == "scope")
+                if (arg == "scopes")
                 {
                     scopes.Clear();
                     scopes.AddRange(args[++i].Split(','));
@@ -64,7 +152,7 @@ namespace netCoreMsal
 
             if (help)
             {
-                Console.WriteLine("// optional arguments --resource --redirectUri --clientId --tenantId --scope --detail");
+                Console.WriteLine("// optional arguments --resource --redirectUri --clientId --clientsecret --tenantId --scope --detail");
                 Console.WriteLine("// run from non administrator prompt!");
                 ShowDetail();
                 return;
@@ -72,46 +160,6 @@ namespace netCoreMsal
 
             if (detail) { ShowDetail(); }
             Authorize();
-        }
-
-        private void Authorize()
-        {
-            AuthenticationResult authenticationResult = default;
-
-            if (string.IsNullOrEmpty(redirectUri))
-            {
-                publicClientApp = PublicClientApplicationBuilder
-                    .Create(clientId)
-                    .WithAuthority(AzureCloudInstance.AzurePublic, tenantId)
-                    .WithDefaultRedirectUri()
-                    .WithLogging(MsalLoggerCallback, LogLevel.Verbose, true, true)
-                    .Build();
-            }
-            else
-            {
-                publicClientApp = PublicClientApplicationBuilder
-                    .CreateWithApplicationOptions(new PublicClientApplicationOptions { ClientId = clientId, RedirectUri = redirectUri })
-                    .WithAuthority(AzureCloudInstance.AzurePublic, tenantId)
-                    .WithLogging(MsalLoggerCallback, LogLevel.Verbose, true, true)
-                    .Build();
-            }
-
-            try
-            {
-                TokenCacheHelper.EnableSerialization(publicClientApp.UserTokenCache);
-                
-                authenticationResult = publicClientApp
-                    .AcquireTokenSilent(scopes, publicClientApp.GetAccountsAsync().Result.FirstOrDefault())
-                    .ExecuteAsync().Result;
-            }
-            catch (MsalUiRequiredException)
-            {
-                authenticationResult = publicClientApp
-                    .AcquireTokenInteractive(scopes)
-                    .ExecuteAsync().Result;
-            }
-
-            FormatJsonOutput(authenticationResult);
         }
 
         private void FormatJsonOutput(AuthenticationResult authenticationResult)
@@ -126,15 +174,15 @@ namespace netCoreMsal
                     writer.WriteString("AccessToken", authenticationResult.AccessToken);
 
                     writer.WriteStartObject("Account");
-                    writer.WriteString("Environment", authenticationResult.Account.Environment);
+                    writer.WriteString("Environment", authenticationResult.Account?.Environment);
 
                     writer.WriteStartObject("HomeAccountId");
-                    writer.WriteString("Identifier", authenticationResult.Account.HomeAccountId.Identifier);
-                    writer.WriteString("ObjectId", authenticationResult.Account.HomeAccountId.ObjectId);
-                    writer.WriteString("TenantId", authenticationResult.Account.HomeAccountId.TenantId);
+                    writer.WriteString("Identifier", authenticationResult.Account?.HomeAccountId.Identifier);
+                    writer.WriteString("ObjectId", authenticationResult.Account?.HomeAccountId.ObjectId);
+                    writer.WriteString("TenantId", authenticationResult.Account?.HomeAccountId.TenantId);
                     writer.WriteEndObject();
 
-                    writer.WriteString("Username", authenticationResult.Account.Username);
+                    writer.WriteString("Username", authenticationResult.Account?.Username);
                     writer.WriteEndObject();
 
                     writer.WriteString("CorrelationId", authenticationResult.CorrelationId);
