--- netCoreMsalTest/netCoreMsal.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <PublishTrimmed>true</PublishTrimmed>
-    <PublishReadyToRun>true</PublishReadyToRun>
-    <PublishSingleFile>true</PublishSingleFile>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-  </PropertyGroup>
-  <ItemGroup>
-    <RuntimeHostConfigurationOption Include="System.Globalization.Invariant" Value="true" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.5.1" />
-    <PackageReference Include="System.Security.Cryptography.ProtectedData" Version="4.7.0-preview1.19504.10" />
-  </ItemGroup>
-</Project>
